Change library name to libindic-stemmer
The current name of library, 'stemmer' is too general and hence
ambiguous to users as to which languages it applies to. The proposed
name "libindic-stemmer" clears this confusion and identifies the stemmer
as a part of the LibIndic project and obviously conveys the fact that it
applies to Indic languages.